Move twing environment setup to a separate file
